+++ src/components/CardList/CardList.tsx
@@ -1,23 +1,17 @@
-import { DetailedCard } from '@/components/DetailedCard';
-import { SimpleCard } from '@/components/SimpleCard';
+import { DetailedCard, SimpleCard } from '@/components';
 import type { CardListProps } from '@/types/interfaces';
 import { isDetailedPokemon } from '@/types/typeGuards';
-import { Component } from 'react';
 
-export class CardList extends Component<CardListProps> {
-  render() {
-    const { results } = this.props;
-
-    return (
-      <div className="flex flex-wrap justify-center  p-4 overflow-auto">
-        {results.map((pokemon) =>
-          isDetailedPokemon(pokemon) ? (
-            <DetailedCard key={pokemon.id} pokemon={pokemon} />
-          ) : (
-            <SimpleCard key={pokemon.name} pokemon={pokemon} />
-          )
-        )}
-      </div>
-    );
-  }
-}
+export const CardList = ({ results }: CardListProps) => {
+  return (
+    <div className="flex flex-wrap justify-center  p-4 overflow-auto">
+      {results.map((pokemon) =>
+        isDetailedPokemon(pokemon) ? (
+          <DetailedCard key={pokemon.id} pokemon={pokemon} />
+        ) : (
+          <SimpleCard key={pokemon.name} pokemon={pokemon} />
+        )
+      )}
+    </div>
+  );
+};
