--- src/components/CardList/index.ts
@@ -1,1 +1,0 @@
-export { CardList } from './CardList';
